+++ app/src/main/java/com/android/solvit/provider/ui/profile/ProviderRegistrationProfile.kt
@@ -103,10 +103,10 @@ import com.android.solvit.shared.model.provider.Language
 import com.android.solvit.shared.model.provider.Provider
 import com.android.solvit.shared.model.service.Services
 import com.android.solvit.shared.model.utils.loadBitmapFromUri
-import com.android.solvit.shared.ui.authentication.GoBackButton
 import com.android.solvit.shared.ui.booking.PackageCard
 import com.android.solvit.shared.ui.navigation.NavigationActions
 import com.android.solvit.shared.ui.utils.CustomOutlinedTextField
+import com.android.solvit.shared.ui.utils.GoBackButton
 import com.android.solvit.shared.ui.utils.ValidationRegex
 
 /**
