+++ app/src/main/java/com/android/solvit/seeker/ui/profile/SeekerRegistrationProfile.kt
@@ -59,9 +59,9 @@ import com.android.solvit.seeker.ui.request.LocationDropdown
 import com.android.solvit.shared.model.authentication.AuthViewModel
 import com.android.solvit.shared.model.map.Location
 import com.android.solvit.shared.model.map.LocationViewModel
-import com.android.solvit.shared.ui.authentication.GoBackButton
 import com.android.solvit.shared.ui.navigation.NavigationActions
 import com.android.solvit.shared.ui.utils.CustomOutlinedTextField
+import com.android.solvit.shared.ui.utils.GoBackButton
 import com.android.solvit.shared.ui.utils.ValidationRegex
 
 /**
