+++ app/src/main/java/com/android/solvit/shared/ui/authentication/ForgotPassword.kt
@@ -45,6 +45,7 @@ import coil.compose.rememberAsyncImagePainter
 import com.android.solvit.R
 import com.android.solvit.shared.ui.navigation.NavigationActions
 import com.android.solvit.shared.ui.utils.CustomOutlinedTextField
+import com.android.solvit.shared.ui.utils.GoBackButton
 import com.android.solvit.shared.ui.utils.ValidationRegex
 
 /**
