+++ app/src/main/java/com/android/solvit/shared/ui/authentication/SignIn.kt
@@ -30,15 +30,12 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.ClickableText
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.Email
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Checkbox
 import androidx.compose.material3.CheckboxDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme.colorScheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
@@ -75,12 +72,12 @@ import com.android.solvit.shared.model.authentication.AuthViewModel
 import com.android.solvit.shared.ui.navigation.NavigationActions
 import com.android.solvit.shared.ui.navigation.Screen
 import com.android.solvit.shared.ui.utils.CustomOutlinedTextField
+import com.android.solvit.shared.ui.utils.GoBackButton
 import com.android.solvit.shared.ui.utils.PasswordTextField
 import com.android.solvit.shared.ui.utils.ValidationRegex
 import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
 import com.google.android.gms.common.api.ApiException
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
 /**
@@ -177,33 +174,6 @@ fun SignInScreen(
           )
         }
       })
-}
-
-/**
- * A composable function that displays a "Go Back" button with a debounce mechanism to prevent
- * multiple rapid clicks.
- *
- * @param navigationActions A set of navigation actions to handle screen transitions.
- */
-@Composable
-fun GoBackButton(navigationActions: NavigationActions) {
-  var canGoBack by remember { mutableStateOf(true) }
-  val coroutineScope = rememberCoroutineScope()
-  IconButton(
-      onClick = {
-        if (canGoBack) {
-          canGoBack = false
-          navigationActions.goBack()
-          coroutineScope.launch {
-            delay(500)
-            canGoBack = true
-          }
-        }
-      },
-      modifier = Modifier.testTag("goBackButton"),
-      enabled = canGoBack) {
-        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "goBackButton")
-      }
 }
 
 /**
