+++ app/src/main/java/com/android/solvit/shared/ui/authentication/SignUpChooseProfile.kt
@@ -49,6 +49,7 @@ import com.android.solvit.seeker.ui.profile.Stepper
 import com.android.solvit.shared.model.authentication.AuthViewModel
 import com.android.solvit.shared.ui.navigation.NavigationActions
 import com.android.solvit.shared.ui.navigation.Screen
+import com.android.solvit.shared.ui.utils.GoBackButton
 
 /**
  * A composable function that displays the "Choose Your Profile" screen during the sign-up process.
